requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
apscheduler>=3.10
lark-oapi>=1.0.0   # 飞书开放平台 SDK（互动卡片 + 长链接审批）
